# openapi.yaml
...
openapi: 3.1.0
info:
  title: VoiceAI engine + platform + telephony
  version: 0.10.217
  description: 'Merged spec: engine (agent CRUD, /health, ws /chat/v1/{agent_id})
    + platform routers (/calls/simulate, /executions/stats, batches, auth, graphs,
    workflows) + telephony (/talko/*, /twilio/*, /plivo/*). Generated from FastAPI
    openapi() (A11 hygiene 2026-09-12).'
paths:
...
  /talko/call:
    post:
      tags:
      - Talko Telephony
      summary: Initiate Outbound Call via Talko (Tata Tele)
      description: Asks talko-service to dial the recipient; Tata streams the answered
        call back to Talko, which relays media to the voiceai agent.
      operationId: make_call_talko_call_post
      parameters:
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TalkoCallDetails'
      responses:
        '200':
          description: Call initiated successfully.
          content:
            application/json:
              schema: {}
        '400':
          description: Missing agent_id / phone / DID / credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Missing or invalid X-API-Key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '502':
          description: talko-service unreachable or rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /talko/hangup:
    post:
      tags:
      - Talko Telephony
      summary: Hang up a Talko call
      description: Passthrough to talko-service hangup for AI-bridge calls.
      operationId: hangup_call_talko_hangup_post
      parameters:
      - name: X-API-Key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TalkoHangupDetails'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '401':
          description: Missing or invalid X-API-Key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '502':
          description: talko-service unreachable or rejected the request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ErrorBody:
      properties:
        code:
          type: string
          title: Code
          description: Stable error code; see voiceai.errors.ErrorCode.
        message:
          type: string
          title: Message
          description: Caller-safe description of what went wrong.
        error_id:
          type: string
          title: Error Id
          description: Correlation id that also appears in the server log.
        retryable:
          type: boolean
          title: Retryable
          description: Whether the same request may succeed if retried.
          default: false
        component:
          anyOf:
          - type: string
          - type: 'null'
          title: Component
          description: Subsystem that failed (llm, synthesizer, api, ...).
        provider:
          anyOf:
          - type: string
          - type: 'null'
          title: Provider
          description: Upstream provider, when the failure was theirs.
        model:
          anyOf:
          - type: string
          - type: 'null'
          title: Model
          description: Model or deployment involved, when known.
        details:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Details
          description: Structured context such as a config path or validation issues.
      type: object
      required:
      - code
      - message
      - error_id
      title: ErrorBody
      description: Structured error as it appears on the wire (``error`` key).
    ErrorEnvelope:
      properties:
        ok:
          type: boolean
          title: Ok
          description: Always false for error bodies.
          default: false
        detail:
          type: string
          title: Detail
          description: Same as error.message; kept for FastAPI-style clients.
        error:
          $ref: '#/components/schemas/ErrorBody'
      type: object
      required:
      - detail
      - error
      title: ErrorEnvelope
      description: Body of every non-2xx HTTP response produced by the engine.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    TalkoCallDetails:
      properties:
        agent_id:
          type: string
          title: Agent Id
          description: voiceai agent id that will handle the call.
        recipient_phone_number:
          type: string
          title: Recipient Phone Number
          description: Customer number to dial (e.g. 919812345678).
        caller_did:
          type: string
          title: Caller Did
          description: Override for the dedicated DID (defaults to TALKO_AI_DID).
          default: ''
        partner_id:
          type: string
          title: Partner Id
          description: Override for TALKO_PARTNER_ID.
          default: ''
        talko_api_key:
          type: string
          title: Talko Api Key
          description: Talko partner API key for this call (defaults to TALKO_API_KEY
            env). Lets each UI user dial with their own key.
          default: ''
      type: object
      required:
      - agent_id
      - recipient_phone_number
      title: TalkoCallDetails
    TalkoHangupDetails:
      properties:
        call_id:
          type: string
          title: Call Id
          description: Vendor call_id to hang up (from /talko/call response when present).
        talko_api_key:
          type: string
          title: Talko Api Key
          description: Talko partner API key (defaults to TALKO_API_KEY env).
          default: ''
      type: object
      required:
      - call_id
      title: TalkoHangupDetails
...